Move detailed tests into their own test module
